// src/api/generated/model/mgmtDistributionSetRequestBodyPost.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtSoftwareModuleAssignment } from './mgmtSoftwareModuleAssignment';

export interface MgmtDistributionSetRequestBodyPost {
  /** The name of the entity */
  name?: string;
  /** The description of the entity */
  description?: string;
  /** Package version */
  version?: string;
  /** Should be set only if change of locked state is requested. If put, the distribution set locked flag will be
set to the requested. Note: unlock (i.e. set this property to false) with extreme care!
In general once distribution set is locked it shall not be unlocked. Note that it could have been assigned /
deployed to targets. */
  locked?: boolean;
  /** True if DS is a required migration step for another DS. As a result the DS’s assignment will not be cancelled
when another DS is assigned (note: updatable only if DS is not yet assigned to a target) */
  requiredMigrationStep?: boolean;
  modules?: MgmtSoftwareModuleAssignment[];
  /** The type of the distribution set */
  type?: string;
}