// src/api/generated/model/mgmtSoftwareModule.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { Links } from './links';

/**
 * **_links**:
* **type** - The software module type of the entity
* **artifacts** - List of artifacts of given software module
* **metadata** - List of metadata

 */
export interface MgmtSoftwareModule {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** The technical identifier of the entity */
  id: number;
  /** Package version */
  version: string;
  /** The software module type of the entity */
  type: string;
  /** The software module type name of the entity */
  typeName?: string;
  /** The software vendor */
  vendor?: string;
  /** If the software module is encrypted */
  encrypted?: boolean;
  /** If the software module is locked */
  locked?: boolean;
  /** If the software module is deleted */
  deleted?: boolean;
  _links?: Links;
}